🐛 Refreshing the knowledge base or Agent configuration page will return to the model configuration page.
