+++ frontend/app/[locale]/setup/agents/components/AgentSetupOrchestrator.tsx
