+++ frontend/app/[locale]/setup/agents/components/DebugConfig.tsx
