+++ frontend/app/[locale]/setup/agents/components/McpConfigModal.tsx
