+++ frontend/app/[locale]/setup/agents/components/PromptManager.tsx
