+++ frontend/app/[locale]/setup/agents/components/agent/AgentConfigModal.tsx
