+++ frontend/app/[locale]/setup/agents/components/agent/CollaborativeAgentDisplay.tsx
