+++ frontend/app/[locale]/setup/agents/components/tool/ToolConfigModal.tsx
