+++ frontend/app/[locale]/setup/agents/components/tool/ToolPool.tsx
