+++ frontend/app/[locale]/setup/agents/config.tsx
