+++ frontend/app/[locale]/setup/agents/page.tsx
@@ -8,9 +8,7 @@ import { App } from "antd";
 import { motion } from "framer-motion";
 
 import { useAuth } from "@/hooks/useAuth";
-import { configService } from "@/services/configService";
 import modelEngineService from "@/services/modelEngineService";
-import { configStore } from "@/lib/config";
 import {
   USER_ROLES,
   CONNECTION_STATUS,
@@ -19,9 +17,9 @@ import {
 import log from "@/lib/logger";
 
 import SetupLayout from "../SetupLayout";
-import DataConfig from "./config";
+import AgentConfig from "./config";
 
-export default function KnowledgeSetupPage() {
+export default function AgentSetupPage() {
   const { message } = App.useApp();
   const router = useRouter();
   const { t } = useTranslation();
@@ -39,33 +37,18 @@ export default function KnowledgeSetupPage() {
       router.push("/");
       return;
     }
+
+    // Only admin users can access this page
+    if (user && user.role !== USER_ROLES.ADMIN) {
+      router.push("/setup/knowledges");
+      return;
+    }
   }, [user, userLoading, router]);
 
   // Check the connection status when the page is initialized
   useEffect(() => {
     checkModelEngineConnection();
-
-    // Trigger knowledge base data acquisition when the page is initialized
-    window.dispatchEvent(
-      new CustomEvent("knowledgeBaseDataUpdated", {
-        detail: { forceRefresh: true },
-      })
-    );
-
-    // Load config for normal user
-    const loadConfigForNormalUser = async () => {
-      if (user && user.role !== USER_ROLES.ADMIN) {
-        try {
-          await configService.loadConfigToFrontend();
-          configStore.reloadFromStorage();
-        } catch (error) {
-          log.error("加载配置失败:", error);
-        }
-      }
-    };
-
-    loadConfigForNormalUser();
-  }, [user]);
+  }, []);
 
   // Function to check the ModelEngine connection status
   const checkModelEngineConnection = async () => {
@@ -82,35 +65,11 @@ export default function KnowledgeSetupPage() {
     }
   };
 
-  // Handle next button click (for admin users)
-  const handleNext = async () => {
-    try {
-      // For admin users, go to agent page
-      router.push("/setup/agent");
-    } catch (error) {
-      log.error("Navigation error:", error);
-      message.error("系统异常，请稍后重试");
-    }
-  };
-
-  // Handle complete button click (for normal users)
+  // Handle complete button click
   const handleComplete = async () => {
     try {
       setIsSaving(true);
-
-      // Reload the config for normal user before saving, ensure the latest model config
-      await configService.loadConfigToFrontend();
-      configStore.reloadFromStorage();
-
-      // Get the current global configuration
-      const currentConfig = configStore.getConfig();
-
-      // Check if the main model is configured
-      if (!currentConfig.models.llm.modelName) {
-        message.error("未找到模型配置，请联系管理员先完成模型配置");
-        return;
-      }
-
+      // Jump to chat page directly, no any check
       router.push("/chat");
     } catch (error) {
       log.error("保存配置异常:", error);
@@ -122,11 +81,7 @@ export default function KnowledgeSetupPage() {
 
   // Handle back button click
   const handleBack = () => {
-    if (user?.role === USER_ROLES.ADMIN) {
-      router.push("/setup/model");
-    } else {
-      message.error(t("setup.page.error.adminOnly"));
-    }
+    router.push("/setup/knowledges");
   };
 
   // Animation variants for smooth transitions
@@ -151,24 +106,18 @@ export default function KnowledgeSetupPage() {
     duration: 0.4,
   };
 
-  // Determine which button to show based on user role
-  const isAdmin = user?.role === USER_ROLES.ADMIN;
-
   return (
     <SetupLayout
       connectionStatus={connectionStatus}
       isCheckingConnection={isCheckingConnection}
       onCheckConnection={checkModelEngineConnection}
       title={t("setup.header.title")}
-      description={t("setup.knowledge.description")}
+      description={t("setup.agent.description")}
       onBack={handleBack}
-      onNext={isAdmin ? handleNext : undefined}
-      onComplete={isAdmin ? undefined : handleComplete}
+      onComplete={handleComplete}
       isSaving={isSaving}
-      showBack={isAdmin}
-      showNext={isAdmin}
-      showComplete={!isAdmin}
-      nextText={t("setup.navigation.button.next")}
+      showBack={true}
+      showComplete={true}
       completeText={t("setup.navigation.button.complete")}
     >
       <motion.div
@@ -179,7 +128,7 @@ export default function KnowledgeSetupPage() {
         transition={pageTransition}
         style={{ width: "100%", height: "100%" }}
       >
-        <DataConfig isActive={true} />
+        <AgentConfig />
       </motion.div>
     </SetupLayout>
   );
