+++ frontend/app/[locale]/setup/knowledges/components/document/DocumentStatus.tsx
