+++ frontend/app/[locale]/setup/knowledges/components/knowledge/KnowledgeBaseList.tsx
