+++ frontend/app/[locale]/setup/knowledges/components/upload/UploadArea.tsx
