+++ frontend/app/[locale]/setup/knowledges/components/upload/UploadAreaUI.tsx
