+++ frontend/app/[locale]/setup/knowledges/config.tsx
