+++ frontend/app/[locale]/setup/models/components/appConfig.tsx
