+++ frontend/app/[locale]/setup/models/components/model/ModelDeleteDialog.tsx
