+++ frontend/app/[locale]/setup/models/components/model/ModelListCard.tsx
