+++ frontend/app/[locale]/setup/models/components/modelConfig.tsx
